refactored oracleUtils into smaller files
