+++ lib/components/oracle/embed/OracleNewDb.tsx
@@ -1,4 +1,3 @@
-import { OracleReportContext } from "@oracle";
 import {
   DropFiles,
   MessageManagerContext,
