+++ lib/components/oracle/reports/tiptap-extensions/OracleReportRecommendationTitle.tsx
@@ -9,7 +9,7 @@ import {
 import React, { useContext, useRef, useState, useEffect } from "react";
 import { Drawer } from "@ui-components";
 import { OracleAnalysisDrawerWithFollowOn } from "../OracleAnalysisDrawerWithFollowOn";
-import { OracleReportContext } from "../../OracleReportContext";
+import { OracleReportContext } from "../../utils/OracleReportContext";
 import { createRoot, Root } from "react-dom/client";
 
 interface RecommendationTitleAttrs {
