+++ lib/components/oracle/reports/tiptap-extensions/comments/OracleCommentHandler.tsx
@@ -1,6 +1,6 @@
 import { useContext, useEffect, useState, useSyncExternalStore } from "react";
 import { OracleCommentPopover } from "./OracleCommentPopover";
-import type { OracleReportComment } from "../../../OracleReportContext";
+import type { OracleReportComment } from "../../../utils/OracleReportContext";
 import {
   Editor,
   mergeAttributes,
@@ -9,9 +9,9 @@ import {
   ReactNodeViewRenderer,
   useCurrentEditor,
 } from "@tiptap/react";
-import { OracleReportContext } from "../../../OracleReportContext";
+import { OracleReportContext } from "../../../utils/OracleReportContext";
 import { MessageSquarePlus } from "lucide-react";
-import type { CommentManager } from "../../../oracleUtils";
+import type { CommentManager } from "../../../utils";
 
 declare module "@tiptap/core" {
   interface Commands<ReturnType> {
