+++ lib/oracle.ts
@@ -3,5 +3,4 @@ export { OracleReport } from "./components/oracle/reports/OracleReport";
 
 export { OracleNav } from "./components/oracle/reports/OracleNav";
 
-export * from "./components/oracle/oracleUtils";
-export * from "./components/oracle/OracleReportContext";
+export * from "./components/oracle/utils";
